Sun, Jan 15, 2023, 4:16 AM +00:00

+++ hours/hours.c
@@ -44,8 +44,8 @@ float calc_hours(int hours[], int weeks, char output)
         {
             sumA += hours[i];
         }
-        sumA = sumA / weeks;
-        return sumA;
+        sumA = sumA / (float) weeks;
+        return (float) sumA;
     }
 
 }
